Rename, re-layer, add build setting
