+++ llvm/lib/Support/raw_ostream.cpp
@@ -619,7 +619,7 @@ raw_fd_ostream::raw_fd_ostream(int fd, bool shouldClose, bool unbuffered,
                                OStreamKind K)
     : raw_pwrite_stream(unbuffered, K), FD(fd), ShouldClose(shouldClose) {
   // FIXME(sandboxing): Remove this by adopting `llvm::vfs::OutputBackend`.
-  auto BypassSandbox = sys::sandbox_scoped_disable();
+  auto BypassSandbox = sys::sandbox::scopedDisable();
 
   if (FD < 0 ) {
     ShouldClose = false;
